Combine the two bounty related action creators
